# app/Resources/api_doc.json
{
    "swagger": "2.0",
    "info": {
        "title": "Page API documentation",
        "version": "1.0.0"
    },
    "host": "rochedor.fr",
    "basePath": "/api",
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/home/{id}": {
            "put": {
                "summary": "Edit requested home",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "sub_title": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "content": {
                                    "type": "object"
                                },
                                "locale": {
                                    "type": "string"
                                },
                                "url": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The home ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "404": {
                        "description": "Page not found"
                    }
                }
            }
        },
        "/home/{id}/versions": {
            "get": {
                "summary": "Return all log entries for the selected home",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "the home id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested logs of your selected home"
                    },
                    "404": {
                        "description": "Home not found"
                    }
                }
            }
        },
...
    },
...
}